new-api: rearrange tests
